# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CleanSpot API - Waste reporting platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /officers:
    get:
      operationId: listOfficers
      tags: [officers]
      summary: List all officers (admin only)
      responses:
        "200":
          description: List of officers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfficerList"
    post:
      operationId: createOfficer
      tags: [officers]
      summary: Create new officer (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOfficerBody"
      responses:
        "201":
          description: Officer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Officer"
  /officers/{id}:
    get:
      operationId: getOfficer
      tags: [officers]
      summary: Get officer details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Officer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Officer"
    patch:
      operationId: updateOfficer
      tags: [officers]
      summary: Update officer (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOfficerBody"
      responses:
        "200":
          description: Updated officer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Officer"
    delete:
      operationId: deleteOfficer
      tags: [officers]
      summary: Delete officer (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /officers/{id}/reports:
    get:
      operationId: getOfficerReports
      tags: [officers]
      summary: Get reports assigned to a specific officer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [reported, cleaning, cleaned]
        - name: limit
          in: query
          schema:
            type: integer
            default: 500
            description: Maximum number of reports to return
        - name: days
          in: query
          schema:
            type: integer
            description: Only return reports from the last N days (e.g. 90)
      responses:
        "200":
          description: Officer's reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportList"
...
components:
  schemas:
    PhotoItem:
      type: object
      properties:
        url:
          type: string
        uploadedAt:
          type: string
          description: ISO-8601 timestamp of when this photo was uploaded
      required: [url, uploadedAt]
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Report:
      type: object
      properties:
        id:
          type: integer
        imageUrl:
          type: string
          nullable: true
        imageUploadedAt:
          type: string
          format: date-time
          nullable: true
          description: Server-recorded timestamp of when the photo was uploaded
        imageUrls:
          type: array
          items:
            $ref: "#/components/schemas/PhotoItem"
          nullable: true
          description: All uploaded report photos with timestamps (max 2)
        cleanupImageUrl:
          type: string
          nullable: true
        cleanupImageUrls:
          type: array
          items:
            $ref: "#/components/schemas/PhotoItem"
          nullable: true
          description: All cleanup photos uploaded by officer (max 5)
        latitude:
          type: number
        longitude:
          type: number
        address:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [reported, cleaning, cleaned]
        reporterEmail:
          type: string
          nullable: true
          description: Only included for admin and panchayat_admin roles
        reporterIp:
          type: string
          nullable: true
        assignedOfficerId:
          type: integer
          nullable: true
        assignedOfficer:
          $ref: "#/components/schemas/OfficerBasic"
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
          description: Set when the report is archived (soft-deleted). Null means the report is active.
        wasteTypes:
          type: array
          items:
            type: string
          nullable: true
          description: AI-identified waste categories (e.g. plastic bottles, food waste)
        brandNames:
          type: array
          items:
            type: string
          nullable: true
          description: AI-identified brand names visible on packaging
        wasteSeverity:
          type: string
          enum: [low, medium, high]
          nullable: true
          description: AI-assessed severity of the waste (low/medium/high)
        photoAiAnalysedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when AI photo analysis was completed
      required: [id, latitude, longitude, status, createdAt, updatedAt]
    ReportList:
      type: object
      properties:
        reports:
          type: array
          items:
            $ref: "#/components/schemas/Report"
        total:
          type: integer
      required: [reports, total]
...
    Officer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        areaName:
          type: string
          nullable: true
        wardName:
          type: string
          nullable: true
          description: Ward assigned to this officer (alias of areaName)
        panchayatName:
          type: string
          nullable: true
        centerLat:
          type: number
          nullable: true
        centerLng:
          type: number
          nullable: true
        reportCount:
          type: integer
        pendingCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, reportCount, pendingCount, createdAt]
    OfficerBasic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        areaName:
          type: string
          nullable: true
        wardName:
          type: string
          nullable: true
          description: Ward assigned to this officer (alias of areaName)
      required: [id, name, email]
    OfficerList:
      type: object
      properties:
        officers:
          type: array
          items:
            $ref: "#/components/schemas/Officer"
        total:
          type: integer
      required: [officers, total]
    CreateOfficerBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        phone:
          type: string
          nullable: true
        areaName:
          type: string
          nullable: true
        panchayatName:
          type: string
          nullable: true
        centerLat:
          type: number
          nullable: true
        centerLng:
          type: number
          nullable: true
      required: [name, email, password]
...
    UpdateOfficerBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
          nullable: true
        password:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        areaName:
          type: string
          nullable: true
        centerLat:
          type: number
          nullable: true
        centerLng:
          type: number
          nullable: true
...